Patch 0529
・MVP獲得者がホムで、主人がMVPアイテムを重量オーバーで所持できないとき、
　nullpoエラーが表示されるバグを修正（mob.c）
	merged from fix by ultramage on AthenaProject, thanks a lot!!
・ポインタへキャストする際のwarningを修正（unit.c）
・0515のdbやconfファイルの空行読み飛ばしにおいて、UNIX環境下で改行コードが
　CR/LFの場合を考慮してなかった問題を修正

+++ src/map/unit.c
@@ -389,7 +389,7 @@ static int unit_walktoxy_timer(int tid,unsigned int tick,int id,void *data)
 				return 0;
 			}
 		}
-		ud->walktimer = add_timer(tick+i,unit_walktoxy_timer,id,(void*)ud->walkpath.path_pos);
+		ud->walktimer = add_timer(tick+i,unit_walktoxy_timer,id,(void*)((int)ud->walkpath.path_pos));
 	} else {
 		// 目的地に着いた
 		if(sd && sd->sc.data[SC_RUN].timer != -1) {
